Creating a new Swift/SceneKit project
Using File -> New -> Project menu in Xcode will allow a new
project to be created. This template was instantiated from the
iOS -> Application -> Game template, with the Swift language,
SceneKit support and targeted to Universal devices (both iPhone
and iPads).

This template was generated from Xcode 6.1 with Swift 1.1, and
later versions of Xcode may vary.

Note that a build error may be seen "Copy SceneKit Assets Error"
for "ship.dae". This does not appear to affect the build and in
any case will be removed in the next step, so don't worry about
it if you see this message.