Adding a white cylinder
A cylinder can be added to the scene by creating a geometry (which
represents the shape) and then adding to a node (which is the item
that gets drawn on screen).

When run this will display a white shadowless cylinder, which
doesn't look like a 3d shape, because it doesn't have any material
or lighting.