Adding lighting and colour
To colour the tree trunk, the colour needs to be set with a
material. This could use a pre-calculated image to give a bark-like
effect, but in this case a solid brown colour will be used from
the UIColor's built-in colour range.

To show, the cylinder needs to have lighting added. Although lights
can be individually added and coloured, using the automatic lighting
provides an easy way to get started.