Adding a cone
The tree's leaves can be represented as a cone, which is added to
the top of the trunk. The leaves colour will be green and is set
as before, but the Y position of the leaves are moved to be at the
top of the trunk itself.

Note that the leaves are added to the trunk node, rather than the
root node. This allows the tree to be moved and the leaves will
automatically follow it to different locations.