Finishing it off
By stacking multiple cones on top of each other, and positioning
them relative to the base of the tree, a tree can be built.

Note that the cones are actually overlapping with each other; the
cone below penetrates the bottom of the cone above. This could be
modified by using a modified version of SCNCone:

  SCNCone(topRadius:1, bottomRadius:3, height:2)

This would result in a closer fit without overlaps, but the top
cone would need to be handled differently. To simplify the code
example used the cropping approach was not used and is left as
an exercise for the reader.